Various improvements.
Rename type to 'Context'; this matches the existing UsePeriodicPasswordProvider.
Add CancellationToken; the user might execute commands against the server that might need to be cancelled.
Remove sync overload; it could be unnecessary.
Move types to their own files.
Allow multiple callbacks to be added.
